+++ src/Exceptions/PageNotAllowedException.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * phpgram
+ *
+ * This File is part of the phpgram Micro Framework
+ *
+ * Web: https://gitlab.com/grammm/php-gram/phpgram
+ *
+ * @license https://gitlab.com/grammm/php-gram/phpgram/blob/master/LICENSE
+ *
+ * @author Jörn Heinemann <joernheinemann@gmx.de>
+ */
+
+namespace Gram\Exceptions;
+
+/**
+ * Class PageNotAllowedException
+ * @package Gram\Exceptions
+ *
+ * Exception für 405
+ */
+class PageNotAllowedException extends AbstractGramException
+{
+
+}