+++ src/Exceptions/PageNotFoundException.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * phpgram
+ *
+ * This File is part of the phpgram Micro Framework
+ *
+ * Web: https://gitlab.com/grammm/php-gram/phpgram
+ *
+ * @license https://gitlab.com/grammm/php-gram/phpgram/blob/master/LICENSE
+ *
+ * @author Jörn Heinemann <joernheinemann@gmx.de>
+ */
+
+namespace Gram\Exceptions;
+
+/**
+ * Class PageNotFoundException
+ * @package Gram\Exceptions
+ *
+ * Wenn 404 Handler nicht gefunden wurden
+ */
+class PageNotFoundException extends AbstractGramException
+{
+
+}